// src/entities/monitoring/crane-status/model/outdoor-work-types.ts
export interface OutdoorCraneRow {
  equipment: string;
  comm: boolean;
  on: boolean;
  fault: boolean;
  notComm: boolean;
  freeSlewing: boolean;
  rotate: boolean;
  trolley1: string;
  trolley2: string;
  gantry: string;
  hoist1: string;
  hoist2: string;
  hoist3: string;
  trolley2Secondary: string;
  slewing: string;
  gantrySecondary: string;
}

export interface OutdoorInfoRow {
  equipment: string;
  equipmentType: string;
  location: string;
  status: string;
  task: string;
  direction: string;
}

export interface OutdoorStatusRow {
  time: string;
  equipment: string;
  statusChange: string;
  level: string;
  location: string;
}